backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
openai==1.12.0
langchain==0.1.9
faiss-cpu==1.7.4
python-dotenv==1.0.1
whisper==1.1.10
pandas==2.2.1
plotly==5.18.0
gunicorn==21.2.0 